+++ src/Accounts/Accounts/Properties/Resources.Designer.cs
@@ -115,7 +115,7 @@ namespace Microsoft.Azure.Commands.Profile.Properties {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to With the new Azure PowerShell login experience, you can select the subscription you want to use more easily. Learn more about it and its configuration at https://go.microsoft.com/fwlink/?linkid=2271236. Share your feedback regarding your experience with `Connect-AzAccount` at: https://aka.ms/azloginfeedback.
+        ///   Looks up a localized string similar to With the new Azure PowerShell login experience, you can select the subscription you want to use more easily. Learn more about it and its configuration at https://go.microsoft.com/fwlink/?linkid=2271909..
         /// </summary>
         internal static string AnnouncementsMessage {
             get {