+++ src/Compute/Compute.Test/ScenarioTests/GalleryTests.cs
@@ -64,5 +64,12 @@ namespace Microsoft.Azure.Commands.Compute.Test.ScenarioTests
         {
             TestRunner.RunTestScript("Test-GalleryVersionWithSourceImageVMId");
         }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestGalleryImageDefinitionDefaults()
+        {
+            TestRunner.RunTestScript("Test-GalleryImageDefinitionDefaults");
+        }
     }
 }